// lvis-web/composables/motorpool/trip-ticket/trip-ticket-approver.types.ts
import type { Employee } from "~/composables/hr/employee/employee.types"


export interface TripTicketApprover {
    id: string
    approver_id: string
    approver: Employee | null
    date_approval: string | null
    notes: string
    status: APPROVAL_STATUS,
    label: string
    order: number
  }


  export interface CreateTripTicketApprover {
    approver: Employee | null
    label: string
    order: number
    showRequiredMsg: boolean
  }


  export interface ChangeApproverResponse {
    success: boolean
    msg: string
    data?: TripTicketApprover
  }